requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
django-cors-headers==4.3.0
psycopg2>=2.8.6,<2.9
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
drf-spectacular>=0.15.1,<0.16
minio==7.2.7
pandas==2.2.2
numpy==1.23.5
s3fs==2024.6.1
fsspec==2024.6.1
openai==1.37.1
cmdstanpy==1.2.4
stanio==0.5.1
prophet==1.1.5
xgboost==2.1.1
scikit-learn==1.5.1
tensorflow==2.12.0
python-dotenv
SQLAlchemy==2.0.34
langchain
langchain-community
langchain-openai